Adding backend validation

+++ Controllers/CommentController.cs
@@ -26,23 +26,29 @@ namespace projeto1.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if(!ModelState.IsValid) return BadRequest(ModelState);
+
             var comments = await _commentRepository.GetAllAsync();
             var commentsDTO = comments.Select(c => c.ToCommentDTO());
             return Ok(commentsDTO);
         }
 
         [HttpGet]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if(!ModelState.IsValid) return BadRequest(ModelState);
+
             var comment = await _commentRepository.GetByIdAsync(id);
             if (comment == null) return NotFound();
             return Ok(comment.ToCommentDTO());
         }
 
-        [HttpPost("{stockId}")]
+        [HttpPost("{stockId:int}")]
         public async Task<IActionResult> Create([FromRoute] int stockId, CreateCommentDTO commentDTO)
         {
+            if(!ModelState.IsValid) return BadRequest(ModelState);
+
             if (!await _stockRepository.StockExistsAsync(stockId)) return BadRequest("Stock does not exist");
 
             var commentModel = commentDTO.ToCommentFromCreate(stockId);
@@ -54,6 +60,8 @@ namespace projeto1.Controllers
         [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDTO updateDTO)
         {
+
+            if(!ModelState.IsValid) return BadRequest(ModelState);
 
             var comment = await _commentRepository.UpdateAsync(id, updateDTO.ToCommentFromUpdate(id));
 
@@ -67,6 +75,8 @@ namespace projeto1.Controllers
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
 
+            if(!ModelState.IsValid) return BadRequest(ModelState);
+            
             var comment = await _commentRepository.DeleteAsync(id);
 
             if (comment == null) return NotFound("Comment does not exist");
